changed return array type

+++ cupy/random/_generator_api.pyx
@@ -305,8 +305,7 @@ class Generator:
             size = (size,)
         elif size is None:
             size = p.shape
-        
-        y = ndarray(size if size is not None else(), numpy.float64)
+        y = ndarray(size if size is not None else(), numpy.int64)
 
         p = cupy.broadcast_to(p, y.shape)
         p_arr = _array_data(p)
